+++ documentation/HISTORY.md
@@ -2,6 +2,10 @@
 
 ## Version 4
 
+## 4.2.2
+- Deprecated css_dir, js_dir, images_dir and swf_dir entries from uri configuration;
+- Removed urlJS, urlCSS, urlIMG e urlSWF template variables.
+
 ## 4.2.1
 - Fixed SendGrid SDK version;
 - Other enhancements and adjustments.